+++ enigma/src/main/java/cuchaz/enigma/analysis/BuiltinPlugin.java
@@ -71,7 +71,7 @@ public final class BuiltinPlugin implements EnigmaPlugin {
 	}
 
 	private void registerDecompilerServices(EnigmaPluginContext ctx) {
-		ctx.registerService("enigma:quiltflower", DecompilerService.TYPE, ctx1 -> Decompilers.QUILTFLOWER);
+		ctx.registerService("enigma:vineflower", DecompilerService.TYPE, ctx1 -> Decompilers.VINEFLOWER);
 		ctx.registerService("enigma:procyon", DecompilerService.TYPE, ctx1 -> Decompilers.PROCYON);
 		ctx.registerService("enigma:cfr", DecompilerService.TYPE, ctx1 -> Decompilers.CFR);
 		ctx.registerService("enigma:bytecode", DecompilerService.TYPE, ctx1 -> Decompilers.BYTECODE);