+++ enigma/src/main/java/cuchaz/enigma/source/Decompilers.java
@@ -3,10 +3,10 @@ package cuchaz.enigma.source;
 import cuchaz.enigma.source.bytecode.BytecodeDecompiler;
 import cuchaz.enigma.source.cfr.CfrDecompiler;
 import cuchaz.enigma.source.procyon.ProcyonDecompiler;
-import cuchaz.enigma.source.quiltflower.QuiltflowerDecompiler;
+import cuchaz.enigma.source.vineflower.VineflowerDecompiler;
 
 public class Decompilers {
-	public static final DecompilerService QUILTFLOWER = QuiltflowerDecompiler::new;
+	public static final DecompilerService VINEFLOWER = VineflowerDecompiler::new;
 	public static final DecompilerService PROCYON = ProcyonDecompiler::new;
 	public static final DecompilerService CFR = CfrDecompiler::new;
 	public static final DecompilerService BYTECODE = BytecodeDecompiler::new;