+++ enigma/src/main/java/cuchaz/enigma/source/vineflower/EnigmaContextSource.java
@@ -1,4 +1,4 @@
-package cuchaz.enigma.source.quiltflower;
+package cuchaz.enigma.source.vineflower;
 
 import cuchaz.enigma.classprovider.ClassProvider;
 import cuchaz.enigma.utils.AsmUtil;