+++ enigma/src/main/java/cuchaz/enigma/source/vineflower/EnigmaTextTokenCollector.java
@@ -1,4 +1,4 @@
-package cuchaz.enigma.source.quiltflower;
+package cuchaz.enigma.source.vineflower;
 
 import cuchaz.enigma.source.SourceIndex;
 import cuchaz.enigma.source.Token;